+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Finalizes CRU(-D) Operations, adding functionality for completing tasks
 - Minimizes repeated code by creating a headers variable and helper functions
 - Migrates setIsSaving instances into helper function
+- Creates a helper function for setting error messages with more user friendly messages
 
 ## [0.0.9] - 2026-01-10